--- skaben/access/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class AlertConfig(AppConfig):
-    name = 'access'
-    verbose_name = 'ALERT: управление уровнями тревоги'
-    default_auto_field = 'django.db.models.BigAutoField'
