+++ skaben/alert/admin.py
@@ -1,5 +1,6 @@
-from core import models
 from django.contrib import admin
 
-admin.site.register(models.AlertCounter)
-admin.site.register(models.AlertState)
+from .models import AlertCounter, AlertState
+
+admin.site.register(AlertCounter)
+admin.site.register(AlertState)
