+++ skaben/assets/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+
+class FileAdmin(admin.ModelAdmin):
+    exclude = ('hash',)
