+++ skaben/assets/models/__init__.py
@@ -1,2 +1,2 @@
-from .files import AudioFile, ImageFile, TextFile, VideoFile
+from .files import AudioFile, File, ImageFile, TextFile, VideoFile
 from .games import HackGame
