+++ skaben/device/views.py
@@ -7,7 +7,7 @@ from rest_framework.decorators import api_view
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
-from .models import Lock, Terminal  # , SimpleLight
+from .models import Lock, Terminal
 
 
 @api_view(http_method_names=['GET'])
@@ -31,12 +31,3 @@ class TerminalViewSet(viewsets.ModelViewSet, DynamicAuthMixin):
     """ Manage terminals in database """
     queryset = Terminal.objects.all()
     serializer_class = serializers.TerminalSerializer
-#
-#
-# class SimpleLightViewSet(viewsets.ModelViewSet):
-#     """ Manage lesser devices in database """
-#     authentication_classes = (TokenAuthentication,)
-#     permission_classes = (IsAuthenticated,)
-#
-#     queryset = SimpleLight.objects.all()
-#     serializer_class = serializers.SimpleLightSerializer
