+++ skaben/shape/models.py
@@ -1,8 +1,6 @@
 from django.conf import settings
 from django.db import models
 
-from .alert import AlertState
-
 
 class MenuItem(models.Model):
 
@@ -33,12 +31,12 @@ class MenuItem(models.Model):
     option = models.CharField(choices=TYPE_CHOICES,
                               default=USER,
                               max_length=32)
-    game = models.ForeignKey("HackGame", on_delete=models.CASCADE, blank=True, null=True)
-    user = models.ForeignKey("UserInput", on_delete=models.CASCADE, blank=True, null=True)
-    audio = models.ForeignKey("AudioFile", on_delete=models.CASCADE, blank=True, null=True)
-    video = models.ForeignKey("VideoFile", on_delete=models.CASCADE, blank=True, null=True)
-    text = models.ForeignKey("TextFile", on_delete=models.CASCADE, blank=True, null=True)
-    image = models.ForeignKey("ImageFile", on_delete=models.CASCADE, blank=True, null=True)
+    game = models.ForeignKey("assets.HackGame", on_delete=models.CASCADE, blank=True, null=True)
+    user = models.ForeignKey("actions.UserInput", on_delete=models.CASCADE, blank=True, null=True)
+    audio = models.ForeignKey("assets.AudioFile", on_delete=models.CASCADE, blank=True, null=True)
+    video = models.ForeignKey("assets.VideoFile", on_delete=models.CASCADE, blank=True, null=True)
+    text = models.ForeignKey("assets.TextFile", on_delete=models.CASCADE, blank=True, null=True)
+    image = models.ForeignKey("assets.ImageFile", on_delete=models.CASCADE, blank=True, null=True)
 
     def __str__(self):
         timer = f"{self.timer}s" if self.timer > 0 else "not set"
@@ -58,10 +56,10 @@ class WorkMode(models.Model):
         verbose_name_plural = 'Настройки: (Терминал) режим работы меню'
 
     name = models.CharField(max_length=48, default="terminal mode")
-    state = models.ManyToManyField(AlertState, blank=True)
+    state = models.ManyToManyField('alert.AlertState', blank=True)
     header = models.CharField(max_length=48, default="terminal vt40k")
     footer = models.CharField(max_length=48, default="unauthorized access is strictly prohibited")
-    menu_set = models.ManyToManyField(MenuItem)
+    menu_set = models.ManyToManyField('shape.MenuItem')
 
     @property
     def url(self):
@@ -81,6 +79,39 @@ class WorkMode(models.Model):
         return f"Mode <{self.name}>"
 
 
+class AccessCode(models.Model):
+    """In-game access code or card id"""
+
+    class Meta:
+        verbose_name = 'Настройки: (Замок) код - данные карты'
+        verbose_name_plural = 'Настройки: (Замок) код - данные карты'
+
+    code = models.CharField(max_length=8)
+    name = models.CharField(max_length=64)
+    surname = models.CharField(max_length=64)
+    position = models.CharField(max_length=128)
+
+    def __str__(self):
+        return f'<{self.code}> {self.position} {self.surname}'
+
+
+class Permission(models.Model):
+    """Access codes permission to open locks"""
+
+    class Meta:
+        unique_together = ('card', 'lock')
+        verbose_name = 'Настройки: (Замок) код - права допуска'
+        verbose_name_plural = 'Настройки: (Замок) код - права допуска'
+
+    card = models.ForeignKey('shape.AccessCode', on_delete=models.CASCADE)
+    lock = models.ForeignKey('device.Lock', on_delete=models.CASCADE)
+    state_id = models.ManyToManyField('alert.AlertState')
+
+    def __str__(self):
+        return f'[ {self.lock.info.upper()} ] {self.card.position} ' \
+               f'{self.card.surname} '
+
+
 class SimpleConfig(models.Model):
     """
         Simple dumb device, such as lights, sirens, rgb-leds
@@ -92,7 +123,7 @@ class SimpleConfig(models.Model):
 
     config = models.JSONField()
     dev_type = models.CharField(max_length=16)
-    state = models.ForeignKey(AlertState,
+    state = models.ForeignKey('alert.AlertState',
                               on_delete=models.SET_NULL,
                               null=True,
                               blank=True)
